+++ R/crumblr.R
@@ -59,6 +59,8 @@ clr <- function(counts, pseudocount = 0.5) {
 #' @param pseudocount added to counts to avoid issues with zeros
 #' @param method \code{"clr"} computes standard centered log ratio and precision weights based on the delta approximation. \code{"clr_2class"} computes the \code{clr()} transform for category \code{i} using 2 classes: 1) counts in category i, and 2) counts _not_ in category i.
 #' @param tau overdispersion parameter for Dirichlet multinomial.  If \code{NULL}, estimate from observed counts.
+#' @param max.ratio regularize estimates of the weights to have a maximum ratio of \code{max.ratio} between the maximum and \code{quant} quantile value
+#' @param quant quantile value used for \code{max.ratio}
 #'
 #' @return  An \code{EList} object with the following components:
 #' \describe{
@@ -70,7 +72,9 @@ clr <- function(counts, pseudocount = 0.5) {
 #' Evalute the centered log ratio (CLR) transform of the count matrix, and the asymptotic theoretical variances of each transformed observation.  The asymptotic normal approximation is increasingly accurate for small overdispersion \eqn{\tau}, large total counts \eqn{C}, and large proportions \eqn{p}, but shows good agreement with the empirical results in most situtations. In practice, it is often reasonable to assume a sufficient number of counts before a variable is included in an analysis anyway.  But the feasability of this assumption is up to the user to determine.
 #'
 #' Given the array \code{p} storing proportions for one sample across all categories, the delta approximation uses the term \code{1/p}.  This can be unstable for small values of \code{p}, and the estimated variances can be sensitive to small changes in the proprtions.  To address this, the \code{"clr_2class"} method computes the \code{clr()} transform for category \code{i} using 2 classes: 1) counts in category i, and 2) counts _not_ in category i. Since class (2) now sums counts across all other categories, the small proportions are avoided and the variance estimates are more stable.
-#'
+#' 
+#' For real data, the asymptotic variance formula can give weights that vary substantially across samples and give very high weights for a subset of samples.  In order to address this, we regularize the weights to reduce the variation in the weights to have a maximum ratio of \code{max.ratio} between the maximum and \code{quant} quantile value.  
+#' 
 #' @examples
 #' # set probability of each category
 #' prob <- c(0.1, 0.2, 0.3, 0.5)
@@ -100,12 +104,11 @@ clr <- function(counts, pseudocount = 0.5) {
 #' fit <- eBayes(fit)
 #'
 #' topTable(fit, coef = "Age", sort.by = "none")
-#' #
 #' @seealso \code{limma::voom}, \code{variancePartition::dream}
 #' @export
 setGeneric(
   "crumblr",
-  function(counts, pseudocount = 0.5, method = c("clr", "clr_2class"), tau = 1) {
+  function(counts, pseudocount = 0.5, method = c("clr", "clr_2class"), tau = 1, max.ratio=10, quant=0.05) {
     standardGeneric("crumblr")
   }
 )
@@ -117,11 +120,11 @@ setGeneric(
 #' @importFrom methods new
 setMethod(
   "crumblr", "matrix",
-  function(counts, pseudocount = 0.5, method = c("clr", "clr_2class"), tau = 1) {
+  function(counts, pseudocount = 0.5, method = c("clr", "clr_2class"), tau = 1, max.ratio=10, quant=0.05) {
     method <- match.arg(method)
 
     if (method == "clr") {
-      res <- .crumblr(counts, pseudocount, tau)
+      res <- .crumblr(counts, pseudocount, tau, max.ratio, quant)
     } else {
       res <- .clr_2class(counts, pseudocount)
     }
@@ -134,20 +137,30 @@ setMethod(
 #' @importFrom methods new
 setMethod(
   "crumblr", "data.frame",
-  function(counts, pseudocount = 0.5, method = c("clr", "clr_2class"), tau = 1) {
+  function(counts, pseudocount = 0.5, method = c("clr", "clr_2class"), tau = 1, max.ratio=10, quant=0.05) {
     method <- match.arg(method)
 
     if (method == "clr") {
-      res <- .crumblr(as.matrix(counts), pseudocount, tau)
+      res <- .crumblr(as.matrix(counts), pseudocount, tau, max.ratio, quant)
     } else {
-      res <- .clr_2class(as.matrix(counts), pseudocount)
+      res <- .clr_2class(as.matrix(counts), pseudocount, tau, max.ratio, quant)
     }
     res
   }
 )
 
+# @param max.ratio regularize estimates of the weights to have a maximum ratio of \code{max.ratio} between the maximum and \code{quant} qauntile value
+# @param quant qauntile value used for \code{max.ratio}
+.cap_ratio = function(W, max.ratio=10, quant=0.05){
+  t(apply(W, 1, function(x){
+    x = x / quantile(x, quant)
+    pmin(x,max.ratio)
+    }))
+}
 
-.crumblr <- function(counts, pseudocount = 0.5, tau = 1) {
+# @param max.ratio regularize estimates of the weights to have a maximum ratio of \code{max.ratio} between the maximum and \code{quant} qauntile value
+# @param quant qauntile value used for \code{max.ratio}
+.crumblr <- function(counts, pseudocount = 0.5, tau = 1, max.ratio=10, quant=0.5) {
   D <- ncol(counts)
 
   # estimate overdispersion from observed counts
@@ -163,11 +176,17 @@ setMethod(
     tau * (1 / p - 2 / (p * D) + sum(1 / p) / D^2) / C
   })
 
+  # regularise the variance estimates
+  weights = 1 / var_asymp
+  weights = .cap_ratio( weights, max.ratio, quant)
+
+  # get CLR values
   Y_clr <- t(clr(counts, pseudocount))
 
   # return clr transformed counts and the corresponding observation level precision
   # (i.e. inverse variances)
-  new("EList", list(E = Y_clr, weights = 1 / var_asymp))
+  new("EList", list(E = Y_clr, 
+                    weights = weights))
 }
 
 
@@ -176,7 +195,7 @@ setMethod(
 # 2) sum of all other counts
 # This avoids the 1/p term for small values of p in the
 # delta approximation for rare cell types
-.clr_2class <- function(counts, pseudocount = 0.5) {
+.clr_2class <- function(counts, pseudocount = 0.5, tau = 1, max.ratio=10, quant=0.05) {
   # sunm counts for each row
   rs <- rowSums(counts)
 
@@ -192,7 +211,7 @@ setMethod(
     colnames(df)[1] <- x
 
     # run clr with crumblr weights on 2 class counts
-    obj <- .crumblr(df, pseudocount)
+    obj <- .crumblr(df, pseudocount, tau, max.ratio, quant)
 
     # extract positive class
     obj[1, ]